refactor: 不要な関数を削除する
