+++ app/[lang]/generation/_components/editor-task-view/generation-task-action-dropdown-menu.tsx
@@ -1,3 +1,4 @@
+import { ThumbnailImageSizeType } from "@/app/[lang]/generation/_types/thumbnail-image-size-type"
 import { Button } from "@/components/ui/button"
 import {
   DropdownMenu,
@@ -14,8 +15,8 @@ import {
 import { MoreHorizontalIcon } from "lucide-react"
 
 type Props = {
-  thumbnailSize: string
-  onChange(size: string): void
+  thumbnailSize: ThumbnailImageSizeType
+  onChange(size: ThumbnailImageSizeType): void
 }
 
 /**
