+++ app/[lang]/generation/_types/thumbnail-image-size-type.ts
@@ -1,0 +1,1 @@
+export type ThumbnailImageSizeType = "small" | "middle" | "big"
