+++ app/[lang]/generation/tasks/[task]/_components/generation-task-view.tsx
@@ -11,20 +11,19 @@ import {
 import { PrivateImage } from "@/app/_components/private-image"
 import { AuthContext } from "@/app/_contexts/auth-context"
 import { AppConfirmDialog } from "@/components/app/app-confirm-dialog"
+import { AppFixedContent } from "@/components/app/app-fixed-content"
 import { AppLoadingPage } from "@/components/app/app-loading-page"
 import { Button } from "@/components/ui/button"
 import { Dialog, DialogContent, DialogTrigger } from "@/components/ui/dialog"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { Separator } from "@/components/ui/separator"
 import { Textarea } from "@/components/ui/textarea"
+import { config } from "@/config"
 import { deleteImageGenerationTaskMutation } from "@/graphql/mutations/delete-image-generation-task"
 import { updateRatingImageGenerationTaskMutation } from "@/graphql/mutations/update-rating-image-generation-task"
 import { imageGenerationTaskQuery } from "@/graphql/queries/image-generation/image-generation-task"
-import { skipToken, useMutation, useSuspenseQuery } from "@apollo/client"
-
-import { AppFixedContent } from "@/components/app/app-fixed-content"
-import { config } from "@/config"
 import { cn } from "@/lib/utils"
+import { skipToken, useMutation, useSuspenseQuery } from "@apollo/client"
 import {
   ArrowDownToLine,
   ArrowUpRightSquare,
